Fix RAM, implement functions to read and write mem
The dp_ram is word-addressed but the core (or at least this
particular instantiation of it) is treating addresses as word
addresses.

This commit changes the RAM to by byte addressed, adds functions
for reading and writing memory from the testbench, and adds the
insertion of some instructions to exercise the memory to the
testbench.

Picorv32 has 4MB of RAM in our version - this commit also matches
that with 4MB of RAM for RI5CY.

ChangeLog:

        * verilator-model/dp_ram.sv: (dp_ram) Change words to
        bytes. Handle all addresses as byte addresses. (readByte,
        writeByte) new function and task.
        * verilator-model/ram.sv: (ram) change to 4MB of RAM.
        * verilator-model/testbench.cpp: (main) Poke a sequence of
        instructions into memory prior to execution start.